Feat & Style: 상품등록리스트 생성 및 주석 설명
