+++ src/App.jsx
@@ -1,10 +1,12 @@
 // import MyProfile from './pages/MyProfile';
+import UserProduct from './features/UserProduct';
 import UserProfile from './features/UserProfile'
 
 function App() {
     return (
         <>
             <UserProfile />
+            <UserProduct />
         </>
     );
 }
